Version.

+++ src/depends/iconv/win_iconv.c
@@ -818,7 +818,7 @@ win_iconv(iconv_t _cd, const char **inbuf, size_t *inbytesleft, char **outbuf, s
     {
         if (outbuf != NULL && *outbuf != NULL && cd->to.flush != NULL)
         {
-            outsize = cd->to.flush(&cd->to, (uchar *)*outbuf, *outbytesleft);
+            outsize = cd->to.flush(&cd->to, (uchar *)*outbuf, (int) *outbytesleft);
             if (outsize == -1)
                 return (size_t)(-1);
             *outbuf += outsize;
@@ -836,7 +836,7 @@ win_iconv(iconv_t _cd, const char **inbuf, size_t *inbytesleft, char **outbuf, s
         mode = cd->from.mode;
         wsize = MB_CHAR_MAX;
 
-        insize = cd->from.mbtowc(&cd->from, (const uchar *)*inbuf, *inbytesleft, wbuf, &wsize);
+        insize = cd->from.mbtowc(&cd->from, (const uchar *)*inbuf, (int) *inbytesleft, wbuf, &wsize);
         if (insize == -1)
             return (size_t)(-1);
 
@@ -881,7 +881,7 @@ win_iconv(iconv_t _cd, const char **inbuf, size_t *inbytesleft, char **outbuf, s
             }
         }
 
-        outsize = cd->to.wctomb(&cd->to, wbuf, wsize, (uchar *)*outbuf, *outbytesleft);
+        outsize = cd->to.wctomb(&cd->to, wbuf, wsize, (uchar *)*outbuf, (int) *outbytesleft);
         if (outsize == -1)
         {
             cd->from.mode = mode;
@@ -1120,7 +1120,7 @@ check_utf_bom(rec_iconv_t *cd, ushort *wbuf, int *wbufsize)
 static char *
 strrstr(const char *str, const char *token)
 {
-    int len = strlen(token);
+    size_t len = strlen(token);
     const char *p = str + strlen(str);
 
     while (str <= --p)
